src/test/java/com/pado/SpringBootrealpractice/MemberRepositoryTest.java
package com.pado.SpringBootrealpractice;

import com.pado.SpringBootrealpractice.domain.Member;
import com.pado.SpringBootrealpractice.repository.MemberRepository;
import org.assertj.core.api.Assertions;
import org.junit.jupiter.api.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.annotation.Rollback;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.transaction.annotation.Transactional;

import static org.junit.jupiter.api.Assertions.*;

@RunWith(SpringRunner.class)
@SpringBootTest
class MemberRepositoryTest {

/*
    @Autowired
    MemberRepository memberRepository;

    @Test
    @Transactional // test에 있으면 db를 롤백시켜버림.
    @Rollback(value = false)
    public void testMember() throws Exception{
        //given
        Member member = new Member();
        member.setUsername("memberA");

        //when
        Long saveId = memberRepository.save(member);
        Member findMember = memberRepository.find(saveId);

        //then
        Assertions.assertThat(findMember.getId()).isEqualTo(member.getId());
        Assertions.assertThat(findMember.getUsername()).isEqualTo(member.getUsername());
        Assertions.assertThat(findMember).isEqualTo(member); // 저장한거랑 조회간게 같음. 같은 트랜젝션 안에서 영속성.


    }
*/

}